RH-37064: Use prCheck instead of GitBear
