--- .coafile_without_fails
@@ -1,3 +1,0 @@
-[GitBear]
-files =
-bears = GitBear
