+++ requirements-dev.txt
@@ -3,7 +3,6 @@
 autopep8==1.3.5  # prevents Coala PEP8Bear incorrect reporting
 coala-bears
 colorama==0.3.9  # prevents library version conflict for Coala
-gitpython
 pytest
 pytest-cov
 SQLAlchemy-Utils
